2068 move validation.js to new directory, create readme and update used refs
